started working on test orion pcs simd full e2e
