+++ poly_commit/src/orion.rs
@@ -7,6 +7,9 @@ pub use utils::{
 mod linear_code;
 pub use linear_code::{OrionCodeParameter, ORION_CODE_PARAMETER_INSTANCE};
 
+#[cfg(test)]
+mod linear_code_tests;
+
 mod base_field_impl;
 pub use base_field_impl::{
     orion_commit_base_field, orion_open_base_field, orion_verify_base_field,
